refactor(services -> adapters): rename `services` crate to `adapters`, update all references, and restructure the crate for modularity and clarity
