+++ crates/adapters/src/lib.rs
