+++ crates/adapters/src/repos/news.rs
