+++ crates/adapters/src/repos/source.rs
@@ -1,18 +1,20 @@
-use crate::utils::coders::Url;
+use crate::utils::codecs::Url;
 use async_trait::async_trait;
 use sqlx::types::chrono::{DateTime, Utc};
 use sqlx::{FromRow, SqlitePool, query, query_as};
+use uninews_core::errors::DomainError;
 use uninews_core::models::source::atom::{AtomDraft, AtomSource};
 use uninews_core::models::source::telegram::{TelegramChannelDraft, TelegramChannelSource};
 use uninews_core::models::source::{SourceType, SourceTypeValue};
-use uninews_core::services::{SourceCreate, SourceService};
+use uninews_core::repos::SourceCreate;
+use uninews_core::repos::source::SourceRepository;
 use uuid::Uuid;
 
-pub struct SqliteSourceService {
+pub struct SqliteSourceRepository {
     db_pool: SqlitePool,
 }
 
-impl SqliteSourceService {
+impl SqliteSourceRepository {
     #[must_use]
     pub const fn new(db_pool: SqlitePool) -> Self {
         Self { db_pool }
@@ -31,7 +33,7 @@ struct SourceQueryResult {
 }
 
 impl TryFrom<SourceQueryResult> for SourceType {
-    type Error = String;
+    type Error = DomainError;
 
     fn try_from(source: SourceQueryResult) -> Result<Self, Self::Error> {
         Ok(match source.source {
@@ -40,25 +42,31 @@ impl TryFrom<SourceQueryResult> for SourceType {
                 source.created_at,
                 source
                     .atom_url
-                    .ok_or_else(|| "Missing URL for Atom source".to_string())?
+                    .ok_or_else(|| {
+                        DomainError::InvalidInput("Missing URL for Atom source".to_string())
+                    })?
                     .0,
             )),
             SourceTypeValue::Telegram => Self::TelegramChannel(TelegramChannelSource::new(
                 source.id,
-                source
-                    .telegram_username
-                    .ok_or_else(|| "Missing username for Telegram source".to_string())?,
+                source.telegram_username.ok_or_else(|| {
+                    DomainError::InvalidInput("Missing username for Telegram source".to_string())
+                })?,
                 source.created_at,
             )),
         })
     }
 }
 
-impl SqliteSourceService {
-    async fn insert_atom(&self, draft: AtomDraft) -> Result<(), String> {
+impl SqliteSourceRepository {
+    async fn insert_atom(&self, draft: AtomDraft) -> Result<(), DomainError> {
         let id = Uuid::now_v7();
 
-        let mut tx = self.db_pool.begin().await.map_err(|e| e.to_string())?;
+        let mut tx = self
+            .db_pool
+            .begin()
+            .await
+            .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         let result = query!(
             r#"
@@ -70,14 +78,16 @@ impl SqliteSourceService {
         )
         .execute(&mut *tx)
         .await
-        .map_err(|e| e.to_string())?;
+        .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         if result.rows_affected() == 0 {
-            tx.rollback().await.map_err(|e| e.to_string())?;
-            return Err(format!(
+            tx.rollback()
+                .await
+                .map_err(|e| DomainError::Internal(e.to_string()))?;
+            return Err(DomainError::Conflict(format!(
                 "[atom_feed={0}] mapping {1} already exists",
                 draft.url, draft.source_id
-            ));
+            )));
         }
 
         query!(
@@ -90,7 +100,7 @@ impl SqliteSourceService {
         )
         .execute(&mut *tx)
         .await
-        .map_err(|e| e.to_string())?;
+        .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         let url = draft.url.as_str();
 
@@ -104,20 +114,33 @@ impl SqliteSourceService {
         )
         .execute(&mut *tx)
         .await
-        .map_err(|e| e.to_string())?;
+        .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         if result.rows_affected() == 0 {
-            tx.rollback().await.map_err(|e| e.to_string())?;
-            return Err(format!("[atom_feed={url}] mapping already exists"));
+            tx.rollback()
+                .await
+                .map_err(|e| DomainError::Internal(e.to_string()))?;
+            return Err(DomainError::Conflict(format!(
+                "[atom_feed={url}] mapping already exists"
+            )));
         }
 
-        tx.commit().await.map_err(|e| e.to_string())?;
+        tx.commit()
+            .await
+            .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         Ok(())
     }
 
-    async fn insert_telegram_channel(&self, draft: TelegramChannelDraft) -> Result<(), String> {
-        let mut tx = self.db_pool.begin().await.map_err(|e| e.to_string())?;
+    async fn insert_telegram_channel(
+        &self,
+        draft: TelegramChannelDraft,
+    ) -> Result<(), DomainError> {
+        let mut tx = self
+            .db_pool
+            .begin()
+            .await
+            .map_err(|e| DomainError::Internal(e.to_string()))?;
         let id = Uuid::now_v7();
 
         let result = query!(
@@ -130,14 +153,16 @@ impl SqliteSourceService {
         )
         .execute(&mut *tx)
         .await
-        .map_err(|e| e.to_string())?;
+        .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         if result.rows_affected() == 0 {
-            tx.rollback().await.map_err(|e| e.to_string())?;
-            return Err(format!(
+            tx.rollback()
+                .await
+                .map_err(|e| DomainError::Internal(e.to_string()))?;
+            return Err(DomainError::Conflict(format!(
                 "[telegram_channel={0}] mapping {1} already exists",
                 draft.username, draft.source_id
-            ));
+            )));
         }
 
         query!(
@@ -150,7 +175,7 @@ impl SqliteSourceService {
         )
         .execute(&mut *tx)
         .await
-        .map_err(|e| e.to_string())?;
+        .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         let result = query!(
             r#"
@@ -163,25 +188,29 @@ impl SqliteSourceService {
         )
         .execute(&mut *tx)
         .await
-        .map_err(|e| e.to_string())?;
+        .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         if result.rows_affected() == 0 {
-            tx.rollback().await.map_err(|e| e.to_string())?;
-            return Err(format!(
+            tx.rollback()
+                .await
+                .map_err(|e| DomainError::Internal(e.to_string()))?;
+            return Err(DomainError::Conflict(format!(
                 "[telegram_channel={0}] username already exists",
                 draft.username
-            ));
+            )));
         }
 
-        tx.commit().await.map_err(|e| e.to_string())?;
+        tx.commit()
+            .await
+            .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         Ok(())
     }
 }
 
 #[async_trait]
-impl SourceService for SqliteSourceService {
-    async fn add(&self, draft: SourceCreate) -> Result<(), String> {
+impl SourceRepository for SqliteSourceRepository {
+    async fn add(&self, draft: SourceCreate) -> Result<(), DomainError> {
         match draft {
             SourceCreate::Atom(draft) => self.insert_atom(draft).await?,
             SourceCreate::TelegramChannel(draft) => self.insert_telegram_channel(draft).await?,
@@ -190,7 +219,7 @@ impl SourceService for SqliteSourceService {
         Ok(())
     }
 
-    async fn get_by_id(&self, id: Uuid) -> Result<SourceType, String> {
+    async fn get_by_id(&self, id: Uuid) -> Result<SourceType, DomainError> {
         query_as!(
             SourceQueryResult,
             r#"
@@ -214,12 +243,12 @@ impl SourceService for SqliteSourceService {
         )
         .fetch_optional(&self.db_pool)
         .await
-        .map_err(|e| e.to_string())?
-        .ok_or_else(|| "Source not found".to_string())?
+        .map_err(|e| DomainError::Internal(e.to_string()))?
+        .ok_or_else(|| DomainError::NotFound("Source not found".to_string()))?
         .try_into()
     }
 
-    async fn get_all(&self) -> Result<Vec<SourceType>, String> {
+    async fn get_all(&self) -> Result<Vec<SourceType>, DomainError> {
         let news = query_as!(
             SourceQueryResult,
             r#"
@@ -240,18 +269,22 @@ impl SourceService for SqliteSourceService {
         )
         .fetch_all(&self.db_pool)
         .await
-        .map_err(|e| e.to_string())?;
+        .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         let sources = news
             .into_iter()
             .map(TryInto::try_into)
-            .collect::<Result<Vec<SourceType>, String>>()?;
+            .collect::<Result<Vec<SourceType>, DomainError>>()?;
 
         Ok(sources)
     }
 
-    async fn delete_by_id(&self, id: Uuid) -> Result<(), String> {
-        let mut tx = self.db_pool.begin().await.map_err(|e| e.to_string())?;
+    async fn delete_by_id(&self, id: Uuid) -> Result<(), DomainError> {
+        let mut tx = self
+            .db_pool
+            .begin()
+            .await
+            .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         query!(
             r#"
@@ -266,9 +299,11 @@ impl SourceService for SqliteSourceService {
         )
         .execute(&mut *tx)
         .await
-        .map_err(|e| e.to_string())?;
+        .map_err(|e| DomainError::Internal(e.to_string()))?;
 
-        tx.commit().await.map_err(|e| e.to_string())?;
+        tx.commit()
+            .await
+            .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         Ok(())
     }
@@ -277,8 +312,12 @@ impl SourceService for SqliteSourceService {
         &self,
         source_id: Uuid,
         source_type: SourceTypeValue,
-    ) -> Result<(), String> {
-        let mut tx = self.db_pool.begin().await.map_err(|e| e.to_string())?;
+    ) -> Result<(), DomainError> {
+        let mut tx = self
+            .db_pool
+            .begin()
+            .await
+            .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         let result = query!(
             r#"
@@ -294,14 +333,20 @@ impl SourceService for SqliteSourceService {
         )
         .execute(&mut *tx)
         .await
-        .map_err(|e| e.to_string())?;
+        .map_err(|e| DomainError::Internal(e.to_string()))?;
 
         if result.rows_affected() == 0 {
-            tx.rollback().await.map_err(|e| e.to_string())?;
-            return Err(format!("source {source_id} not found"));
+            tx.rollback()
+                .await
+                .map_err(|e| DomainError::Internal(e.to_string()))?;
+            return Err(DomainError::NotFound(format!(
+                "source {source_id} not found"
+            )));
         }
 
-        tx.commit().await.map_err(|e| e.to_string())?;
+        tx.commit()
+            .await
+            .map_err(|e| DomainError::Internal(e.to_string()))?;
         Ok(())
     }
 }
