+++ crates/adapters/src/services.rs
