+++ crates/adapters/src/services/http.rs
