+++ crates/adapters/src/services/news.rs
