+++ crates/adapters/src/services/storage.rs
