+++ crates/adapters/src/utils.rs
@@ -1,4 +1,4 @@
-pub mod coders;
+pub mod codecs;
 pub mod fs;
 pub mod parse;
 pub mod uuid;
