+++ crates/adapters/src/utils/codecs.rs
