+++ crates/adapters/src/utils/fs.rs
