+++ crates/adapters/src/utils/uuid.rs
