+++ crates/cli/Cargo.toml
@@ -15,9 +15,9 @@ name = "uninews"
 path = "src/main.rs"
 
 [dependencies]
+uninews_adapters = { path = "../adapters" }
 uninews_collect = { path = "../collect" }
 uninews_core = { path = "../core" }
-uninews_services = { path = "../services" }
 
 clap.workspace = true
 dotenvy.workspace = true
