+++ crates/cli/src/commands/source/remove/telegram.rs
@@ -1,10 +1,10 @@
 use clap::Args;
 use std::error::Error;
 use std::sync::Arc;
+use uninews_adapters::utils::parse::parse_telegram_username;
 use uninews_core::models::source::SourceTypeValue;
 use uninews_core::models::source::telegram::TelegramChannelDraft;
 use uninews_core::repos::source::SourceRepository;
-use uninews_services::utils::parse::parse_telegram_username;
 
 #[derive(Debug, Args)]
 pub struct RemoveTelegramChannel {
