+++ crates/collect/Cargo.toml
@@ -11,8 +11,8 @@ repository.workspace = true
 version.workspace = true
 
 [dependencies]
+uninews_adapters = { path = "../adapters" }
 uninews_core = { path = "../core" }
-uninews_services = { path = "../services" }
 
 ammonia = { version = "^4.1.2" }
 async-trait.workspace = true
