+++ crates/collect/src/source/telegram/update_handler.rs
@@ -5,11 +5,11 @@ use async_trait::async_trait;
 use futures::future::{try_join, try_join_all};
 use scraper::{Html, Selector};
 use std::sync::Arc;
+use uninews_adapters::utils::parse::truncate_with_dots;
 use uninews_core::models::News;
 use uninews_core::models::source::telegram::TelegramChannelSource;
 use uninews_core::services::{HttpResponse, HttpUpdateHandler};
 use uninews_core::uuid::gen_consistent_uuid;
-use uninews_services::utils::parse::truncate_with_dots;
 use url::Url;
 
 #[allow(dead_code)]
