+++ crates/core/Cargo.toml
@@ -1,14 +1,14 @@
 [package]
-edition.workspace = true
 name = "uninews_core"
-version.workspace = true
 
 authors.workspace = true
 description.workspace = true
+edition.workspace = true
 homepage.workspace = true
 license.workspace = true
 publish.workspace = true
 repository.workspace = true
+version.workspace = true
 
 [dependencies]
 async-trait.workspace = true
